Minor changes related to structure

+++ .eslintrc.js
@@ -18,7 +18,7 @@ module.exports = {
     'no-console': 1,
     'no-case-declarations': 1,
     'linebreak-style': [2, 'unix'],
-    quotes: [1, 'double'],
+    quotes: [1, 'single'],
     semi: [2, 'always'],
     'prettier/prettier': [2, null, '@prettier']
   }
